+++ examples/with_sensors.rs
@@ -6,8 +6,8 @@ use spaceapi_server::SpaceapiServerBuilder;
 fn main() {
     env_logger::init();
 
-    // Create new minimal Status instance
-    let status = api::StatusBuilder::new("coredump")
+    // Create new minimal Status instance compatible with v0.13 and v14
+    let status = api::StatusBuilder::mixed("coredump")
         .logo("https://www.coredump.ch/logo.png")
         .url("https://www.coredump.ch/")
         .location(api::Location {
@@ -22,6 +22,7 @@ fn main() {
         })
         .add_issue_report_channel(api::IssueReportChannel::Email)
         .add_issue_report_channel(api::IssueReportChannel::Twitter)
+        .state(api::State::default())
         .build()
         .expect("Creating status failed");